Revert "DFPL-2305 Create urgent approveOrder judiciary task - Upload additional applications"
