+++ ccd-definition/AuthorisationCaseField/CareSupervision/cafcass-solicitor.json
@@ -373,13 +373,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseFieldID": "skipDraftOrderUrgencyPage",
-    "UserRole": "[CAFCASSSOLICITOR]",
-    "CRUD": "CRU"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "CaseFieldID": "c2ApplicationType",
     "UserRole": "[CAFCASSSOLICITOR]",
     "CRUD": "CRU"
