--- ccd-definition/CaseEventToComplexTypes/upload-c2/DraftOrderUrgency.json
@@ -1,20 +1,0 @@
-[
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "DraftOrderUrgency",
-    "CaseEventID": "uploadAdditionalApplications",
-    "CaseFieldID": "draftOrderUrgency",
-    "ListElementCode": "urgency",
-    "FieldDisplayOrder": "1",
-    "DisplayContext": "OPTIONAL"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "ID": "DraftOrderUrgency",
-    "CaseEventID": "uploadAdditionalApplications",
-    "CaseFieldID": "draftOrderUrgency",
-    "ListElementCode": "urgencyLabel",
-    "FieldDisplayOrder": "2",
-    "DisplayContext": "READONLY"
-  }
-]
