+++ ccd-definition/CaseEventToFields/CareSupervision/uploadAdditionalApplications.json
@@ -126,34 +126,8 @@
     "PageDisplayOrder": 3,
     "PageColumnNumber": 1,
     "ShowSummaryChangeOption": "Y",
-    "PageShowCondition": "additionalApplicationType CONTAINS \"OTHER_ORDER\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "uploadAdditionalApplications",
-    "CaseFieldID": "skipDraftOrderUrgencyPage",
-    "PageFieldDisplayOrder": 1,
-    "DisplayContext": "READONLY",
-    "PageID": 4,
-    "PageDisplayOrder": 4,
-    "PageColumnNumber": 1,
-    "ShowSummaryChangeOption": "Y",
-    "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/upload-additional-applications/populate-data/mid-event",
-    "FieldShowCondition": "additionalApplicationType=\"DO_NOT_SHOW\""
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "CaseEventID": "uploadAdditionalApplications",
-    "CaseFieldID": "draftOrderUrgency",
-    "PageFieldDisplayOrder": 2,
-    "DisplayContext": "COMPLEX",
-    "PageID": 4,
-    "PageDisplayOrder": 4,
-    "PageColumnNumber": 1,
-    "ShowSummaryChangeOption": "Y",
-    "PageShowCondition": "skipDraftOrderUrgencyPage!=\"Yes\""
+    "PageShowCondition": "additionalApplicationType CONTAINS \"OTHER_ORDER\"",
+    "CallBackURLMidEvent": "${CCD_DEF_CASE_SERVICE_BASE_URL}/callback/upload-additional-applications/populate-data/mid-event"
   },
   {
     "LiveFrom": "01/01/2017",
