+++ ccd-definition/CaseField/CareSupervision/caseField.json
@@ -1034,16 +1034,6 @@
   {
     "LiveFrom": "01/01/2017",
     "CaseTypeID": "CARE_SUPERVISION_EPO",
-    "ID": "skipDraftOrderUrgencyPage",
-    "Label": "Skip draft order urgency page",
-    "FieldType": "YesOrNo",
-    "SecurityClassification": "Public",
-    "Searchable": "N",
-    "Comment": "Hidden field used to control whether we skip the draft order urgency page (i.e. for additional app)"
-  },
-  {
-    "LiveFrom": "01/01/2017",
-    "CaseTypeID": "CARE_SUPERVISION_EPO",
     "ID": "c2ApplicationType",
     "Label": "Upload C2",
     "FieldType": "C2ApplicationTypeSelection",
