+++ service/src/integrationTest/java/uk/gov/hmcts/reform/fpl/controllers/UploadAdditionalApplicationsControllerAboutToStartTest.java
@@ -5,9 +5,7 @@ import org.springframework.boot.test.autoconfigure.OverrideAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
-import uk.gov.hmcts.reform.fpl.enums.YesNo;
 import uk.gov.hmcts.reform.fpl.model.CaseData;
-import uk.gov.hmcts.reform.fpl.model.DraftOrderUrgencyOption;
 import uk.gov.hmcts.reform.fpl.model.Other;
 import uk.gov.hmcts.reform.fpl.model.Others;
 import uk.gov.hmcts.reform.fpl.model.Respondent;
@@ -28,17 +26,6 @@ class UploadAdditionalApplicationsControllerAboutToStartTest extends AbstractCal
 
     UploadAdditionalApplicationsControllerAboutToStartTest() {
         super("upload-additional-applications");
-    }
-
-    @Test
-    void shouldClearDraftOrderUrgency() {
-        CaseData caseData = CaseData.builder()
-            .draftOrderUrgency(DraftOrderUrgencyOption.builder().urgency(List.of(YesNo.YES)).build())
-            .build();
-
-        AboutToStartOrSubmitCallbackResponse response = postAboutToStartEvent(caseData);
-
-        assertThat(response.getData().get("draftOrderUrgency")).isNull();
     }
 
     @Test
