+++ service/src/main/java/uk/gov/hmcts/reform/fpl/enums/WorkAllocationTaskType.java
@@ -5,6 +5,5 @@ public enum WorkAllocationTaskType {
     ORDER_NOT_UPLOADED,
     ORDER_UPLOADED,
     CORRESPONDENCE_UPLOADED,
-    CMO_REVIEWED,
-    DRAFT_ORDER_UPLOADED_WITH_C2
+    CMO_REVIEWED
 }
